pyproject.toml
[project]
name = "werewolf"
version = "0.1.0"
description = "线下狼人杀助手服务"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["server"]
